+++ packages/terra-list/src/List.jsx
@@ -58,10 +58,17 @@ const List = ({
   if (role && role.length > 0 && role !== 'none') {
     attrSpread.role = role;
   }
+  const allItems = children ? [] : undefined;
 
+  React.Children.forEach(children, (child) => {
+    const cloneChild = React.cloneElement(child, {
+      listrole: role,
+    });
+    allItems.push(cloneChild);
+  });
   return (
     <ul {...customProps} {...attrSpread} className={listClassNames} ref={refCallback}>
-      {children}
+      {allItems}
     </ul>
   );
 };