+++ packages/terra-list/src/terra-dev-site/doc/example/ListSubsection.jsx
@@ -23,7 +23,7 @@ const ListSubsectionExample = () => (
       </Item>
     </Subsection>
     <Subsection
-      key="collapsible-subsection"
+      key="collapsible-subsection-1"
       isCollapsible
       title="Collapsible Subsection"
     >
@@ -38,7 +38,7 @@ const ListSubsectionExample = () => (
       </Item>
     </Subsection>
     <Subsection
-      key="collapsible-subsection"
+      key="collapsible-subsection-2"
       isCollapsed
       isCollapsible
       title="Collapsed Subsection"