+++ packages/terra-list/CHANGELOG.md
@@ -3,6 +3,8 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Accessibility issue fixed to handle selectable and non-selectable items correctly.
 
 4.21.0 - (October 14, 2019)
 ------------------